% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui.R
\name{app_ui}
\alias{app_ui}
\title{Application User Interface}
\usage{
app_ui(request)
}
\arguments{
\item{request}{Internal parameter for {shiny}. DO NOT REMOVE.}
}
\description{
Defines the overall UI layout for the Shiny application including
navigation bars, menus, and loading UI modules.
}
