% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_links.R
\name{blast_uniprot}
\alias{blast_uniprot}
\title{BLAST UniProt Search Function}
\usage{
blast_uniprot(
  input,
  db_swiss,
  db_trembl,
  top_n = 1,
  tmp_fasta = "query_tmp.fasta",
  blastp_path
)
}
\arguments{
\item{input}{Character vector of protein sequences, list of sequences,
or path to FASTA file}

\item{db_swiss}{Character string path to SwissProt BLAST database}

\item{db_trembl}{Character string path to TrEMBL BLAST database}

\item{top_n}{Integer number of top hits to return (default: 1)}

\item{tmp_fasta}{Character string temporary FASTA file name (default: "query_tmp.fasta")}

\item{blastp_path}{Character string path to BLASTp executable}
}
\value{
A data frame containing BLAST results with columns: SeqID,
  UniProtAcc, EntryName, Description, Identity, Database
}
\description{
Performs BLASTp search against SwissProt and TrEMBL databases to find
matching UniProt entries for protein sequences.
}
\examples{
\dontrun{
result <- blast_uniprot(
  input = "MAEGEITTFTALTEKFNLPPGNYKKPKLLYCSNGGHFLRILPDGTVDGTRDRSDQHIQLQLSAESIGEG",
  db_swiss = "/path/to/uniprot_sprot",
  db_trembl = "/path/to/uniprot_trembl",
  blastp_path = "/path/to/blastp"
)
}
}
