% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co_enrichment.R
\name{co_enrichment_server}
\alias{co_enrichment_server}
\title{Co-enrichment Visualization Server Module}
\usage{
co_enrichment_server(id)
}
\arguments{
\item{id}{A unique module id.}
}
\value{
Internal reactive objects only.
}
\description{
Server logic for the ProtVis co-enrichment visualization module.
}
