% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co_enrichment.R
\name{co_enrichment_ui}
\alias{co_enrichment_ui}
\title{Co-enrichment Visualization UI Module}
\usage{
co_enrichment_ui(id)
}
\arguments{
\item{id}{A unique module id.}
}
\value{
A Shiny UI component.
}
\description{
Publication-ready ProtVis module for visualizing shared and specific
enrichment patterns across two-omics or three-omics datasets.
}
