% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_chord.R
\name{correlation_chord_server}
\alias{correlation_chord_server}
\title{Correlation Chord Diagram Server
Server module for correlation chord diagram visualization.}
\usage{
correlation_chord_server(id)
}
\arguments{
\item{id}{Module id.}
}
\value{
Shiny server module.
}
\description{
Correlation Chord Diagram Server
Server module for correlation chord diagram visualization.
}
