% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_chord.R
\name{correlation_chord_ui}
\alias{correlation_chord_ui}
\title{Correlation Chord Diagram UI
UI module for correlation chord diagram visualization.}
\usage{
correlation_chord_ui(id)
}
\arguments{
\item{id}{Module id.}
}
\value{
Shiny UI.
}
\description{
Correlation Chord Diagram UI
UI module for correlation chord diagram visualization.
}
