% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input.R
\name{data_input_ui}
\alias{data_input_ui}
\title{Data Input UI Module}
\usage{
data_input_ui(id)
}
\arguments{
\item{id}{Module ID for namespacing.}
}
\value{
A Shiny UI tag list.
}
\description{
This UI module dynamically displays a header showing the current data source
and renders the corresponding UI for the selected data source module.
}
