% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG.R
\name{DEG_server}
\alias{DEG_server}
\title{DEG Analysis Server Logic}
\arguments{
\item{id}{A unique ID for the Shiny module, used to create input/output bindings.}
}
\description{
This function contains the server-side logic for performing DEG (Differential Expression Analysis), including PCA and volcano plot generation, and DEG result calculations.
}
