% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG.R
\name{DEG_ui}
\alias{DEG_ui}
\title{DEG Analysis UI}
\usage{
DEG_ui(id)
}
\arguments{
\item{id}{A unique ID for the Shiny module, used to create input/output bindings.}
}
\description{
This function creates the user interface for DEG (Differential Expression Analysis) that includes file upload options, PCA plot settings, and volcano plot settings.
}
