% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEP_visualize.R
\name{DEP_visualize_server}
\alias{DEP_visualize_server}
\title{DEP Visualization Server Logic}
\arguments{
\item{input}{Shiny input values.}

\item{output}{Shiny output values.}

\item{session}{Shiny session object.}
}
\description{
Defines the server-side logic for the DEP visualization section in a Shiny
application, including plot generation based on user input.
}
