% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DR_analysis.R
\name{DR_analysis_server}
\alias{DR_analysis_server}
\title{DR_analysis_server}
\arguments{
\item{id}{The module ID. This is used to namespace the inputs and outputs in the UI and server components.}
}
\description{
Server logic for Dimension reduction analysis
}
\details{
Dimension reduction analysis Server Module
}
