% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DR_analysis.R
\name{DR_analysis_ui}
\alias{DR_analysis_ui}
\title{DR_analysis_ui}
\usage{
DR_analysis_ui(id)
}
\arguments{
\item{id}{A unique identifier for the Shiny namespace, Dimension reduction analysis.}
}
\description{
Dimension reduction analysis UI Module
}
\details{
Dimension reduction analysis UI Module
}
