% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVolumes_win.R
\name{getVolumes_win}
\alias{getVolumes_win}
\title{Cross-Platform Guide: Checking and Repairing Disk Volumes}
\usage{
getVolumes_win(exclude = NULL)
}
\arguments{
\item{exclude}{A vector of names to exclude from the returned list.}
}
\value{
A named vector of available volumes with optional exclusions.
}
\description{
This function retrieves available volume information based on the operating system
the R session is currently running on. It supports MacOS, Linux, and Windows.
Specifically for Windows, it addresses a bug in `shinyFiles::getVolumes` related
to handling paths with Chinese characters, ensuring correct handling of such paths.
}
