% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO_and_KEGG.R
\name{GO_and_KEGG_server}
\alias{GO_and_KEGG_server}
\title{GO and KEGG Server Logic}
\usage{
GO_and_KEGG_server(id)
}
\arguments{
\item{id}{A unique identifier for the Shiny namespace.}
}
\description{
Implements the logic for GO and KEGG analysis.
}
