% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO_and_KEGG.R
\name{GO_and_KEGG_ui}
\alias{GO_and_KEGG_ui}
\title{GO_and_KEGG_ui}
\usage{
GO_and_KEGG_ui(id)
}
\arguments{
\item{id}{A unique identifier for the Shiny namespace.}
}
\description{
GO and KEGG Database User Interface
}
\details{
Creates a user interface for accessing GO (Gene Ontology) and KEGG (Kyoto Encyclopedia of Genes and Genomes) databases
in a Shiny application.
}
