% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui.R
\name{golem_add_external_resources}
\alias{golem_add_external_resources}
\title{Add External Resources to the Shiny Application}
\usage{
golem_add_external_resources()
}
\description{
This internal function adds external resources such as CSS, JS,
and favicon to the Shiny app. It also sets up resource paths
for static files within the `app/www` directory.
}
