% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv_imputation.R
\name{mv_imputation_ui}
\alias{mv_imputation_ui}
\title{mv_imputation_ui}
\usage{
mv_imputation_ui(id)
}
\arguments{
\item{id}{A unique identifier for the Shiny namespace, Missing value filling.}
}
\description{
Missing value filling UI Module
}
\details{
Missing value filling UI Module
}
