% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv_noise.R
\name{mv_noise_ui}
\alias{mv_noise_ui}
\title{mv_noise_ui}
\usage{
mv_noise_ui(id)
}
\arguments{
\item{id}{A unique identifier for the Shiny namespace}
}
\description{
Remove noise from proteomics data
}
\details{
Remove noise data UI Module
}
