% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv_summary.R
\name{mv_summary_ui}
\alias{mv_summary_ui}
\title{mv_summary_ui}
\usage{
mv_summary_ui(id)
}
\arguments{
\item{id}{A unique identifier for the Shiny namespace.}
}
\description{
Multivariate Summary UI Module
}
\details{
Multivariate Summary UI Module
}
