% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenMS.R
\name{OpenMS_ui}
\alias{OpenMS_ui}
\title{OpenMS User Interface Module
This function creates the user interface for processing proteomics data
from OpenMS output files.}
\usage{
OpenMS_ui(id)
}
\arguments{
\item{id}{Character. Module ID used for namespacing the UI elements.}
}
\value{
UI layout for the OpenMS processing module.
}
\description{
OpenMS User Interface Module
This function creates the user interface for processing proteomics data
from OpenMS output files.
}
