% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview.R
\name{overview_ui}
\alias{overview_ui}
\title{Overview UI Module
Creates the user interface for the overview analysis module}
\usage{
overview_ui(id)
}
\arguments{
\item{id}{Character string specifying the namespace id}
}
\value{
A Shiny UI tagList containing the overview analysis interface
}
\description{
Overview UI Module
Creates the user interface for the overview analysis module
}
