% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_strict_module.R
\name{pd_default_config}
\alias{pd_default_config}
\title{Create default config for the PD strict engine}
\usage{
pd_default_config()
}
\value{
A named list of default engine parameters.
}
\description{
Create default config for the PD strict engine
}
