% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_strict_module.R
\name{pd_strict_module_server}
\alias{pd_strict_module_server}
\title{PD strict spectrum annotation server module}
\usage{
pd_strict_module_server(id)
}
\arguments{
\item{id}{Shiny module id.}
}
\value{
A list containing a reactive `result`.
}
\description{
PD strict spectrum annotation server module
}
