% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_strict_module.R
\name{pd_strict_module_ui}
\alias{pd_strict_module_ui}
\title{PD strict spectrum annotation UI module}
\usage{
pd_strict_module_ui(id)
}
\arguments{
\item{id}{Shiny module id.}
}
\value{
A UI definition for the module.
}
\description{
PD strict spectrum annotation UI module
}
