% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_strict_module.R
\name{pd_strict_protvis_module_final}
\alias{pd_strict_protvis_module_final}
\title{PD strict ProtVis module final single-file version}
\description{
This file contains:
- strict PTM-aware PD spectrum engine helpers
- ProtVis-style Shiny module UI/server
- a standalone demo app runner
}
\details{
It is designed to be placed directly in the `R/` directory of ProtVis.
}
