% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PD.R
\name{PD_ui}
\alias{PD_ui}
\title{PD User Interface Module
This function creates the user interface for processing proteomics data
from Proteome Discoverer (PD) output files.}
\usage{
PD_ui(id)
}
\arguments{
\item{id}{Character. Module ID used for namespacing the UI elements.}
}
\value{
UI layout for the PD processing module.
}
\description{
PD User Interface Module
This function creates the user interface for processing proteomics data
from Proteome Discoverer (PD) output files.
}
