% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\name{plot_enrichment_bar}
\alias{plot_enrichment_bar}
\title{plot_enrichment_bar}
\usage{
plot_enrichment_bar(
  enrich_df,
  top_n = 10,
  fill_color = "#2c7bb6",
  title = NULL
)
}
\description{
plot_enrichment_bar
}
\keyword{internal}
