% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\name{plot_enrichment_dot}
\alias{plot_enrichment_dot}
\title{plot_enrichment_dot}
\usage{
plot_enrichment_dot(
  enrich_df,
  top_n = 10,
  point_color = "#2c7bb6",
  title = NULL
)
}
\description{
plot_enrichment_dot
}
\keyword{internal}
