% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_init_ui.R
\name{project_init_server}
\alias{project_init_server}
\title{Project Initialization Server Module}
\usage{
project_init_server(id, shared_state)
}
\arguments{
\item{id}{Module ID}

\item{shared_state}{A reactiveValues object for sharing state (workdir, sample info, etc.)}
}
\description{
Project Initialization Server Module
}
