% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_init_ui.R
\name{project_init_ui}
\alias{project_init_ui}
\title{Project Initialization UI Module}
\usage{
project_init_ui(id)
}
\arguments{
\item{id}{Module ID}
}
\value{
UI elements for project initialization including working directory setup,
  sample info and expression matrix uploads, data source selection, and data preview.
}
\description{
Project Initialization UI Module
}
