% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_structure.R
\name{protein_structure_ui}
\alias{protein_structure_ui}
\title{Protein Structure UI}
\usage{
protein_structure_ui(id)
}
\arguments{
\item{id}{A unique identifier for the Shiny namespace}
}
\description{
Protein Structure Analysis UI Module
Creates a user interface for protein secondary structure analysis using
bio3d package to perform Normal Mode Analysis (NMA) on PDB files.
}
