% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Raw.R
\name{Raw_server}
\alias{Raw_server}
\title{Raw Data Upload Module Server}
\usage{
Raw_server(id)
}
\arguments{
\item{id}{A character string representing the Shiny module namespace ID.}
}
\value{
A reactive expression that returns a list containing:
}
\description{
Handles the server-side logic for raw data uploads. It renders file
metadata (name, size, and extension) to the UI and returns a reactive
object containing the file path and name for use in other modules.
}
