% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Raw.R
\name{Raw_ui}
\alias{Raw_ui}
\title{Raw Data Upload Module UI}
\usage{
Raw_ui(id)
}
\arguments{
\item{id}{A character string representing the Shiny module namespace ID.}
}
\value{
A Shiny UI tag list containing a file input and a text output.
}
\description{
Provides a simple user interface for uploading raw data files. It includes
a file input selector and a display area to show basic metadata about the
uploaded file.
}
