% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_strict_module.R
\name{run_pd_strict_demo_app}
\alias{run_pd_strict_demo_app}
\title{Run the PD strict module as a standalone demo app}
\usage{
run_pd_strict_demo_app()
}
\value{
A Shiny app object.
}
\description{
Run the PD strict module as a standalone demo app
}
