% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_ProtVis}
\alias{run_ProtVis}
\title{Run the Shiny Application}
\usage{
run_ProtVis(
  onStart = NULL,
  options = list(),
  enableBookmarking = NULL,
  uiPattern = "/",
  progress = getOption("protvis.startup.progress", TRUE),
  ...
)
}
\arguments{
\item{onStart}{A function that will be called before the app is actually run.}

\item{options}{Named list of values passed to `shiny::shinyOptions`.}

\item{enableBookmarking}{Can be "url", "server", or "disable".}

\item{uiPattern}{A regular expression used to determine which requests should be handled by the UI.}

\item{progress}{Logical; show a compact console progress bar while ProtVis initializes.}

\item{...}{Reserved for backward compatibility; currently ignored.}
}
\value{
An object that represents the app.
}
\description{
Launches the ProtVis Shiny application for proteomics data analysis.
}
