% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skyline.R
\name{skyline_ui}
\alias{skyline_ui}
\title{Skyline User Interface Module
This function creates the user interface for processing proteomics data
from Skyline output files.}
\usage{
skyline_ui(id)
}
\arguments{
\item{id}{Character. Module ID used for namespacing the UI elements.}
}
\value{
UI layout for the Skyline processing module.
}
\description{
Skyline User Interface Module
This function creates the user interface for processing proteomics data
from Skyline output files.
}
