% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringdb_ppi.R
\name{stringdb_ppi_server}
\alias{stringdb_ppi_server}
\title{STRINGdb PPI Network Server}
\usage{
stringdb_ppi_server(id)
}
\arguments{
\item{id}{A unique module ID.}
}
\value{
A Shiny server module.
}
\description{
STRINGdb PPI Network Server
}
