% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringdb_ppi.R
\name{stringdb_ppi_ui}
\alias{stringdb_ppi_ui}
\title{STRINGdb PPI Network UI}
\usage{
stringdb_ppi_ui(id)
}
\arguments{
\item{id}{A unique module ID.}
}
\value{
A Shiny UI.
}
\description{
STRINGdb PPI Network UI
}
