% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMT.R
\name{TMT_ui}
\alias{TMT_ui}
\title{TMT User Interface
Creates a user interface for the TMT (Tandem Mass Tag) analysis section
in a Shiny application. This includes file upload, parameter selection,
and visualization options.}
\usage{
TMT_ui(id)
}
\arguments{
\item{id}{A unique identifier for the Shiny namespace.}
}
\description{
TMT User Interface
Creates a user interface for the TMT (Tandem Mass Tag) analysis section
in a Shiny application. This includes file upload, parameter selection,
and visualization options.
}
