% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veen.R
\name{veen_ui}
\alias{veen_ui}
\title{veen_ui}
\usage{
veen_ui(id)
}
\arguments{
\item{id}{A unique identifier for the Shiny namespace, Veen plot}
}
\description{
Veen plot UI Module
}
\details{
Veen plot UI Module
}
