% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgcna.R
\name{wgcna_module}
\alias{wgcna_module}
\title{WGCNA Module for ProtVis}
\description{
Weighted gene co-expression network analysis module adapted for ProtVis.
This module integrates data filtering, soft-threshold selection, module
detection, module-trait association, interested module exploration,
hub-gene extraction, and Cytoscape export.
}
