% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgcna.R
\name{wgcna_server}
\alias{wgcna_server}
\title{WGCNA Server Module}
\usage{
wgcna_server(id, rv = NULL)
}
\arguments{
\item{id}{A unique module id.}

\item{rv}{Optional shared reactiveValues object from ProtVis.}
}
\value{
No return value. This function is called for its side effects.
}
\description{
WGCNA Server Module
}
