% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background_make.R
\name{background_make_ui}
\alias{background_make_ui}
\title{Background Maker Module UI}
\usage{
background_make_ui(id)
}
\arguments{
\item{id}{Character string specifying the namespace id for the module.}
}
\value{
A shiny tagList.
}
\description{
Creates the user interface for the background generation module
that processes EggNOG output files to create GO and KEGG backgrounds.
}
