% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_module.R
\name{boxplot_module_server}
\alias{boxplot_module_server}
\title{Boxplot Module Server}
\usage{
boxplot_module_server(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\value{
A list containing reactive plot object
}
\description{
Server logic for the Boxplot module.
Handles file upload, data processing, long-format conversion,
dynamic color input generation, multi-group comparison,
plotting, and PDF download.
}
