% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_module.R
\name{boxplot_module_ui}
\alias{boxplot_module_ui}
\title{Boxplot Module UI}
\usage{
boxplot_module_ui(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\value{
Shiny UI for boxplot module
}
\description{
Boxplot Module UI
}
