% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEP_analysis.R
\name{DEP_analysis_server}
\alias{DEP_analysis_server}
\title{DEP Analysis Server Module}
\arguments{
\item{id}{The namespace identifier for the module}

\item{shared_state}{A reactive list containing shared state variables across modules}
}
\value{
A reactive list containing comparison data, normalized matrix, sample info, and DEP results
}
\description{
DEP Analysis Server Module
}
