% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEP_visualize.R
\name{DEP_visualize_ui}
\alias{DEP_visualize_ui}
\title{DEP_visualize_ui}
\usage{
DEP_visualize_ui(id)
}
\arguments{
\item{id}{A unique identifier for the Shiny namespace.}
}
\description{
DEP Visualization User Interface
}
\details{
Creates a user interface for visualizing DEP (Differentially Expressed Proteins)
in a Shiny application.
}
