% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DR_analysis.R
\name{DR_analysis_server}
\alias{DR_analysis_server}
\title{DR_analysis_server}
\arguments{
\item{id}{The module ID. This is used to namespace the inputs and outputs in the UI and server components.}

\item{input}{Standard Shiny server arguments: a list of input values from the UI.}

\item{output}{Standard Shiny server arguments: a list of outputs to send back to the UI.}

\item{session}{Standard Shiny server arguments: the session object that maintains the state of the app.}
}
\description{
Server logic for Dimension reduction analysis
}
\details{
Dimension reduction analysis Server Module
}
