% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\name{enrichment_analysis_ui}
\alias{enrichment_analysis_ui}
\title{Enrichment Analysis Module UI}
\arguments{
\item{id}{The namespace identifier for the module}
}
\value{
A Shiny UI tagList containing the enrichment analysis interface
}
\description{
This function creates the user interface for the enrichment analysis module.
}
