% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Expression_profile.R
\name{Expression_profile_ui}
\alias{Expression_profile_ui}
\title{Expression_profile_ui}
\usage{
Expression_profile_ui(id)
}
\arguments{
\item{id}{A unique identifier for the Shiny namespace.}
}
\description{
Expression Profile User Interface
}
\details{
Creates a user interface for displaying expression profiles in a Shiny application.
}
