% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nine_quadrant.R
\name{nine_quadrant_server}
\alias{nine_quadrant_server}
\title{Implements the server-side logic for the Nine-Quadrant Plot module.
This includes reading uploaded files, dynamically generating UI for numeric column selection,
previewing uploaded data, classifying points into quadrants, customizing colors,
rendering the plot, and providing PDF download.}
\arguments{
\item{id}{A character string specifying the namespace of the module.}
}
\value{
A Shiny module server that manages the Nine-Quadrant Plot interactivity.
}
\description{
Implements the server-side logic for the Nine-Quadrant Plot module.
This includes reading uploaded files, dynamically generating UI for numeric column selection,
previewing uploaded data, classifying points into quadrants, customizing colors,
rendering the plot, and providing PDF download.
}
