% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_extract.R
\name{protein_extract_server}
\alias{protein_extract_server}
\title{Protein Extract Server Module
Server for protein sequence extraction from FASTA by protein IDs.}
\usage{
protein_extract_server(id)
}
\arguments{
\item{id}{Module id}
}
\value{
Reactive list
}
\description{
Protein Extract Server Module
Server for protein sequence extraction from FASTA by protein IDs.
}
