% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_extract.R
\name{protein_extract_ui}
\alias{protein_extract_ui}
\title{Protein Extract UI Module
UI for protein sequence extraction from FASTA by protein IDs.}
\usage{
protein_extract_ui(id)
}
\arguments{
\item{id}{Module id}
}
\value{
Shiny UI
}
\description{
Protein Extract UI Module
UI for protein sequence extraction from FASTA by protein IDs.
}
