% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_structure.R
\name{protein_structure_server}
\alias{protein_structure_server}
\title{protein_structure Server Module}
\usage{
protein_structure_server(id)
}
\arguments{
\item{id}{Standard shiny server arguments}
}
\description{
Server logic for protein_structure
}
