% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_structure.R
\name{protein_structure_ui}
\alias{protein_structure_ui}
\title{protein_structure_ui}
\usage{
protein_structure_ui(id)
}
\arguments{
\item{id}{A unique identifier for the Shiny namespace.}
}
\description{
protein_structure
}
