% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_ProtVis}
\alias{run_ProtVis}
\title{Run the Shiny Application}
\usage{
run_ProtVis(
  onStart = NULL,
  options = list(),
  enableBookmarking = NULL,
  uiPattern = "/",
  ...
)
}
\arguments{
\item{onStart}{A function that will be called before the app is actually run.}

\item{options}{Named list of values passed to `shiny::shinyOptions`.}

\item{enableBookmarking}{Can be "url", "server", or "disable".}

\item{uiPattern}{A regular expression used to determine which requests should be handled by the UI.}

\item{...}{Arguments to pass to `golem_opts`. See `?golem::get_golem_options` for more details.}
}
\value{
An object that represents the app.
}
\description{
Launches the ProtVis Shiny application for proteomics data analysis.
}
