% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_normalization.R
\name{sample_subtract}
\alias{sample_subtract}
\title{Perform median subtraction normalization on a data matrix}
\usage{
sample_subtract(data)
}
\arguments{
\item{data}{A numeric matrix or data frame containing the data to be normalized}
}
\value{
A matrix with the same dimensions as input where each column has been median-centered
}
\description{
This function subtracts the median value (ignoring NAs) from each column of the input data matrix.
It is commonly used for sample normalization in omics data analysis.
}
\examples{
data <- matrix(rnorm(100), ncol=5)
normalized <- sample_subtract(data)

}
