% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacked_column_chart.R
\name{stacked_column_chart_server}
\alias{stacked_column_chart_server}
\title{Stacked Bar Chart Server Module}
\usage{
stacked_column_chart_server(id)
}
\arguments{
\item{id}{A unique identifier for the module.}
}
\description{
This server function processes data, generates the stacked bar chart, and handles dynamic plot updates,
custom labels, and download functionality.
}
