% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacked_column_chart.R
\name{stacked_column_chart_ui}
\alias{stacked_column_chart_ui}
\title{Stacked Bar Chart UI Module}
\usage{
stacked_column_chart_ui(id)
}
\arguments{
\item{id}{A unique identifier for the module.}
}
\description{
This UI function creates a stacked bar chart interface with a file upload, color input options,
label customization, and chart download capabilities.
}
