% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMT.R
\name{TMT_server}
\alias{TMT_server}
\title{TMT Server Logic}
\arguments{
\item{input}{Shiny input values.}

\item{output}{Shiny output values.}

\item{session}{Shiny session object.}
}
\description{
Defines the server-side logic for the TMT (Tandem Mass Tag) analysis section
in a Shiny application, including plot generation based on user input.
}
