% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venn.R
\name{venn_ui}
\alias{venn_ui}
\title{venn_ui}
\usage{
venn_ui(id)
}
\arguments{
\item{id}{A unique identifier for the Shiny namespace, venn plot}
}
\description{
venn plot UI Module
}
\details{
venn plot UI Module
}
